grimoire gui, typesetting fixed, auto-generating chapters

+++ main/java/mort/mortmagic/common/grimoire/GrimoirePage.java
@@ -22,7 +22,7 @@ public class GrimoirePage extends IForgeRegistryEntry.Impl<GrimoirePage> impleme
     }
 
     @Override
-    public String getTranslatedText(HashMap<GrimoirePage, Integer> knownPages) {
+    public String getTranslatedText(HashMap<GrimoirePage, Byte> knownPages) {
         return I18n.format( getUntranslatedText( knownPages.get(this) ) );
     }
 }
