+++ main/java/mort/mortmagic/common/grimoire/ICanBeDisplayedInGrimoire.java
@@ -5,6 +5,6 @@ import java.util.Set;
 
 public interface ICanBeDisplayedInGrimoire {
 
-    public String getTranslatedText( HashMap<GrimoirePage,Integer> knownPages );
+    public String getTranslatedText( HashMap<GrimoirePage,Byte> knownPages );
 
 }
